CANT LOG IN
Website fix, migrate login.lua, add a few other things and fixes.

+++ www/guilds.php
@@ -110,7 +110,7 @@ if (user_logged_in() === true) {
 				</select>
 				<input type="text" name="guild_name">
 				
-				<input type="submit" value="Create Guild">
+				<input type="submit" value="Create Guild" class="btn btn-primary">
 			</li>
 		</ul>
 	</form>
@@ -194,12 +194,12 @@ if (user_logged_in() === true) {
 
 <?php if ($inv_count > 0) { ?>
 <h3>Invited characters</h3>
-<table>
+<table id="guildViewTable" class="table table-striped">
 	<tr class="yellow">
-		<td>Name:</td>
+		<th>Name:</th>
 		<?php 
 		if ($highest_access == 2 || $highest_access == 3) {
-			echo '<td>Remove:</td>';
+			echo '<th>Remove:</th>';
 		}
 		// Shuffle through visitor characters
 		for ($i = 0; $i < $char_count; $i++) {
@@ -210,7 +210,7 @@ if (user_logged_in() === true) {
 					$exist = true;
 				}
 			}
-			if ($exist) echo '<td>Join Guild:</td><td>Reject Invitation:</td>';
+			if ($exist) echo '<th>Join Guild:</th><th>Reject Invitation:</th>';
 		}
 		?>
 	</tr>
@@ -219,14 +219,13 @@ if (user_logged_in() === true) {
 		if ($inv_data !== false) foreach ($inv_data as $inv) {
 			$uninv = user_character_data($inv['player_id'], 'name');
 			echo '<tr>';
-			echo '<td>'. $uninv['name'] .'</td>';
+			echo '<td><a href="characterprofile.php?name='. $uninv['name'] .'">'. $uninv['name'] .'</a></td>';
 			// Remove invitation
 			if ($highest_access == 2 || $highest_access == 3) {
-			?> <form action="" method="post"> <?php
+			?> <form action="" method="post" onsubmit="return confirm('Are you sure you want to remove this invitation?')";> <?php
 				echo '<td>';
 				echo '<input type="hidden" name="uninvite" value="' . $inv['player_id'] . '" />';
-				echo '<input type="submit" value="Remove Invitation">';
-				echo '</td>';
+				echo '<input type="submit" value="Remove Invitation" class="btn btn-danger needconfirmation"></td>';
 			?> </form> <?php
 			}
 			// Join Guild
@@ -235,8 +234,7 @@ if (user_logged_in() === true) {
 					if (user_character_id($characters[$i]) == $inv['player_id']) {
 						echo '<td>';
 						echo '<input type="hidden" name="joinguild" value="' . $inv['player_id'] . '" />';
-						echo '<input type="submit" value="Join Guild">';
-						echo '</td>';
+						echo '<input type="submit" value="Join Guild" class="btn btn-primary"></td>';
 						$bool = true;
 					}
 				}
@@ -251,8 +249,7 @@ if (user_logged_in() === true) {
 					if (user_character_id($characters[$i]) == $inv['player_id']) {
 						echo '<td>';
 						echo '<input type="hidden" name="uninvite" value="' . $inv['player_id'] . '" />';
-						echo '<input type="submit" value="Reject Invitation">';
-						echo '</td>';
+						echo '<input type="submit" value="Reject Invitation" class="btn btn-primary"></td>';
 						$bool = true;
 					}
 				}
@@ -278,6 +275,7 @@ if (user_logged_in() === true) {
 		header('Location: guilds.php?name='. $_GET['name']);
 		exit();
 	}
+	
 	if (!empty($_POST['joinguild'])) {
 		// 
 		foreach ($inv_data as $inv) {
@@ -352,6 +350,52 @@ if ($highest_access >= 2) {
 		} else echo '<font color="red" size="4">That character name does not exist.</font>';
 	}
 	
+	if (!empty($_POST['war_invite'])) {
+		$n = get_guild_id($_POST['war_invite']);
+		if (user_guild_exist($n)) {
+			$wars = mysql_select_multi("SELECT `id`, `guild1`, `guild2`, `status` FROM `guild_wars` WHERE (`guild1` = ".$gid." OR `guild1` = ".$n.") AND (`guild2` = ".get_guild_id($_GET['name'])." OR `guild2` = ".$n.") AND (`status` = 0 OR `status` = 1);");		
+			$status = false; $t = guild_war_invite_check($_GET['name']); $v = get_guild_name($gid);
+			if ($t !== false) { 
+				foreach ($t as $x) { 
+					if ($x['id'] == $n) $status = true;
+				} 
+			}
+			
+			foreach ($v as $k) { 
+				if ($k['name'] == $_POST['war_invite']) $status = true;
+			}
+						
+			if ($wars == false && $status == false) {
+				mysql_insert("INSERT INTO `guild_wars` (`guild1`, `guild2`, `name1`, `name2`, `status`, `started`, `ended`) VALUES ('".$gid."', '".$n."', '".$_GET['name']."', '".$_POST['war_invite']."', '0', '".time()."', '0');");
+				echo 'You have invited <a href="guilds.php?name='.$_POST['war_invite'].'">' .$_POST['war_invite'].'</a> to war.';
+			} else echo '<font color="red" size="4">This guild has already been invited to war(or you\'re trying to invite your own guild).</FONT>';
+		} else echo '<font color="red" size="4">That guild name does not exist.</font>';
+	}
+	
+	if (!empty($_POST['war_rdeclaration'])) {
+		guild_war_rdeclaration($_POST['war_rdeclaration'], $gid);
+		header('Location: guilds.php?name='. $_GET['name']);
+		exit();
+	}
+	
+	if (!empty($_POST['war_reject'])) {
+		guild_war_reject($_POST['war_reject'], $gid);
+		header('Location: guilds.php?name='. $_GET['name']);
+		exit();
+	}
+	
+	if (!empty($_POST['war_accept'])) {
+		guild_war_accept($_POST['war_accept'], $gid);
+		header('Location: guilds.php?name='. $_GET['name']);
+		exit();
+	}
+	
+	if (!empty($_POST['war_cancel'])) {
+		guild_war_cancel($_POST['war_cancel'], $gid);
+		header('Location: guilds.php?name='. $_GET['name']);
+		exit();
+	}
+		
 	if (!empty($_POST['disband'])) {
 		// 
 		$gidd = (int)$_POST['disband'];
@@ -469,7 +513,8 @@ if ($highest_access >= 2) {
 						<ul>
 							<li>Create forum guild board:<br>
 							<input type="hidden" name="forumGuildId" value="<?php echo $gid; ?>">
-							<input type="submit" value="Create Guild Board">
+							<input type="submit" value="Create Guild Board" class="btn btn-primary">
+							</button>
 						</ul>
 					</form>
 					<?php
@@ -482,7 +527,7 @@ if ($highest_access >= 2) {
 			<ul>
 				<li>Invite Character to guild:<br>
 					<input type="text" name="invite" placeholder="Character name">
-					<input type="submit" value="Invite Character">
+					<input type="submit" value="Invite Character" class="btn btn-primary">
 				</li>
 			</ul>
 		</form>
@@ -519,7 +564,7 @@ if ($highest_access >= 2) {
 						}
 						?>
 					</select>
-					<input type="submit" value="Promote Member">
+					<input type="submit" value="Promote Member" class="btn btn-primary">
 				</li>
 			</ul>
 		</form>
@@ -542,7 +587,7 @@ if ($highest_access >= 2) {
 					}
 					?>
 					</select>
-					<input type="submit" value="Remove member">
+					<input type="submit" value="Remove member" class="btn btn-danger needconfirmation">
 				</li>
 			</ul>
 		</form>
@@ -560,16 +605,16 @@ if ($highest_access >= 2) {
 						}
 						echo '<input type="hidden" name="change_ranks" value="' . $gid . '" />';
 					?>
-					<input type="submit" value="Update Ranks">
+					<input type="submit" value="Update Ranks" class="btn btn-primary">
 				</li>
 			</ul>
 		</form>
 		<!-- forms to disband guild -->
-		<form action="" method="post">
+		<form action="" method="post" onsubmit="return confirm('Are you sure you want to disband your guild?')";>
 			<ul>
 				<li><b>DELETE GUILD (All members must be offline):</b><br>
 					<?php echo '<input type="hidden" name="disband" value="' . $gid . '" />'; ?>
-					<input type="submit" value="Disband Guild">
+					<input type="submit" value="Disband Guild" class="btn btn-danger needconfirmation">
 				</li>
 			</ul>
 		</form>
@@ -590,19 +635,70 @@ if ($highest_access >= 2) {
 					}
 					?>
 					</select>
-					<input type="submit" value="Change Leadership">
+					<input type="submit" value="Change Leadership" class="btn btn-primary">
 				</li>
 			</ul>
 		</form>
-		<?php }} ?>
-		<?php
+		<?php } ?>
+<?php if ($config['guildwar_enabled'] === true) { ?>
+			<h2>Guild War Management:</h2>
+		<!-- Invite guild to war -->
+		<form action="" method="post">
+			<ul>
+				<li>Invite guild to war:<br>
+					<input type="text" name="war_invite" placeholder="Guild name">
+					<input type="submit" value="Invite Guild" class="btn btn-primary">
+				</li>
+			</ul>
+		</form>
+		
+<!-- cancel war declaration -->
+<?php 
+$t = mysql_select_multi("SELECT `id`, `guild1`, `guild2`, `name1`, `name2`, `status`, `started`, `ended` FROM `guild_wars` WHERE `guild1` = '".$gid."' AND `status` = 0 ORDER BY `started` DESC");
+if (!empty($t)) {
+	echo '<h3>War declarations</h3><table id="guildsTable" class="table table-striped table-hover"><tr class="yellow"><th>#</th><th>Guild Name:</th><th>Remove Invitation</th></tr>';
+	$i = 0;
+		foreach($t as $v) {
+	$i++;
+		echo '<tr><td>'.$i.'</td><td><a href="guilds.php?name='.$v['name2'].'">'.$v['name2'].'</a></td><td><form action="" method="post" onsubmit="return confirm(\'Are you sure you want to cancel your invitation?\')";><input type="hidden" name="war_rdeclaration" value="'.$v['guild2'].'" /><input type="submit" value="Remove Invitation" class="btn btn-danger needconfirmation"></form></td></tr>';
 	}
-}
 ?>
-<!-- end leader-->
+</table>
+<?php } ?>
 
+<!-- accept/reject invitation -->
+<?php 
+$t = mysql_select_multi("SELECT `id`, `guild1`, `guild2`, `name1`, `name2`, `status`, `started`, `ended` FROM `guild_wars` WHERE `guild2` = '".$gid."' AND `status` = 0 ORDER BY `started` DESC");
+if (!empty($t)) {
+	echo '<h3>Pending invitations</h3><table id="guildsTable" class="table table-striped table-hover"><tr class="yellow"><th>#</th><th>Guild Name:</th><th>Accept Invitation</th><th>Reject Invitation</th></tr>';
+	$i = 0;
+		foreach($t as $v) {
+	$i++;
+		echo '<tr><td>'.$i.'</td><td><a href="guilds.php?name='.$v['name1'].'">'.$v['name1'].'</a></td><td><form action="" method="post"><input type="hidden" name="war_accept" value="'.$v['guild1'].'" /><input type="submit" value="Accept Invitation" class="btn btn-primary"></form></td><td><form action="" method="post" onsubmit="return confirm(\'Are you sure you want to reject this invitation?\')";><input type="hidden" name="war_reject" value="'.$v['guild1'].'" /><input type="submit" value="Reject Invitation" class="btn"></form></td></tr>';
+	}
+?>
+</table>
+<?php } ?>
+
+<!-- end war -->
+<?php 
+$t = mysql_select_multi("SELECT `id`, `guild1`, `guild2`, `name1`, `name2`, `status`, `started`, `ended` FROM `guild_wars` WHERE (`guild1` = ".$gid." OR `guild2` = ".$gid.") AND `status` = 1 ORDER BY `started` DESC");
+if (!empty($t)) {
+	echo '<h3>Active Guild Wars</h3><table id="guildsTable" class="table table-striped table-hover"><tr class="yellow"><th>#</th><th>Guild Name:</th><th>Cancel War</th></tr>';
+	$i = 0;
+		foreach($t as $v) {
+	$i++;
+	if ($v['guild1'] == $gid) {
+		echo '<tr><td>'.$i.'</td><td><a href="guilds.php?name='.$v['name2'].'">'.$v['name2'].'</a></td><td><form action="" method="post" onsubmit="return confirm(\'Are you sure you want to cancel this war?\')";><input type="hidden" name="war_cancel" value="'.$v['guild2'].'" /><input type="submit" value="Cancel War" class="btn btn-danger needconfirmation"></form></td></tr>';
+	} else echo '<tr><td>'.$i.'</td><td><a href="guilds.php?name='.$v['name1'].'">'.$v['name1'].'</a></td><td><form action="" method="post" onsubmit="return confirm(\'Are you sure you want to cancel this war?\')";><input type="hidden" name="war_cancel" value="'.$v['guild1'].'" /><input type="submit" value="Cancel War" class="btn btn-danger needconfirmation"></form></td></tr>';
+	}
+?>
+</table>
+<?php } ?>
+<?php } } } ?>
+<!-- end leader-->
+<?php } ?>
 <?php
-/*
 if ($config['TFSVersion'] == 'TFS_02' || $config['TFSVersion'] == 'TFS_10') $wardata = get_guild_wars();
 else if ($config['TFSVersion'] == 'TFS_03') $wardata = get_guild_wars03();
 else die("Can't recognize TFS version. It has to be either TFS_02 or TFS_03. Correct this in config.php");
@@ -615,12 +711,12 @@ if ($wardata !== false) {
 if ($war_exist && $config['guildwar_enabled'] === true) {
 ?>
 <h2>War overview:</h2>
-<table>
+<table id="guildsTable" class="table table-striped table-hover">
 	<tr class="yellow">
-		<td>Attacker:</td>
-		<td>Defender:</td>
-		<td>status:</td>
-		<td>started:</td>
+		<th>Attacker:</th>
+		<th>Defender:</th>
+		<th>status:</th>
+		<th>started:</th>
 	</tr>
 		<?php
 		foreach ($wardata as $wars) {
@@ -636,7 +732,7 @@ if ($war_exist && $config['guildwar_enabled'] === true) {
 		}
 		?>
 </table>
-<?php } */ ?> 
+<?php } ?>
 <!-- leave guild with character -->
 <?php
 $bool = false;
@@ -671,7 +767,7 @@ if ($forumExist !== false) {
 			}
 			?>
 			</select>
-			<input type="submit" value="Leave Guild">
+			<input type="submit" value="Leave Guild" class="btn btn-primary">
 		</li>
 	</ul>
 </form>
